% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash.R
\name{trash}
\alias{trash}
\title{March Madness Trash Talk Generator}
\usage{
trash(x)
}
\arguments{
\item{x}{A string of any length describing the team you wanto to trash.}
}
\value{
A string containing a trash talk one liner.
}
\description{
March Madness Trash Talk Generator
}

